refactor ClientSideRedirectExceptions (#29737)
