+++ components/rest/RestReferencePage.tsx
@@ -1,6 +1,4 @@
-import React, { useState, useEffect } from 'react'
-import { useRouter } from 'next/router'
-import dynamic from 'next/dynamic'
+import React, { useEffect } from 'react'
 import cx from 'classnames'
 
 import { DefaultLayout } from 'components/DefaultLayout'
@@ -11,47 +9,14 @@ import styles from './RestOperation.module.scss'
 import { useAutomatedPageContext } from 'components/context/AutomatedPageContext'
 import { Operation } from './types'
 import { ClientSideHighlight } from 'components/ClientSideHighlight'
-
-const ClientSideRedirectExceptions = dynamic(
-  () => import('components/article/ClientsideRedirectExceptions'),
-  {
-    ssr: false,
-  }
-)
+import { ClientSideRedirects } from 'components/ClientSideRedirects'
 
 export type StructuredContentT = {
   restOperations: Operation[]
 }
 
 export const RestReferencePage = ({ restOperations }: StructuredContentT) => {
-  const { asPath } = useRouter()
   const { title, intro, renderedPage } = useAutomatedPageContext()
-  // We have some one-off redirects for rest api docs
-  // currently those are limited to the repos page, but
-  // that will grow soon as we restructure the rest api docs.
-  // This is a workaround to updating the hardcoded links
-  // directly in the REST API code in a separate repo, which
-  // requires many file changes and teams to sign off.
-  // While the organization is turbulent, we can do this.
-  // Once it's more settled, we can refactor the rest api code
-  // to leverage the OpenAPI urls rather than hardcoded urls.
-  // The code below determines if we should bother loading this redirecting
-  // component at all.
-  // The reason this isn't done at the server-level is because there you
-  // can't possibly access the URL hash. That's only known in client-side
-  // code.
-  const [loadClientsideRedirectExceptions, setLoadClientsideRedirectExceptions] = useState(false)
-  useEffect(() => {
-    const { hash } = window.location
-
-    // Today, Jan 2022, it's known explicitly what the pathname.
-    // In the future there might be more.
-    // Hopefully, we can some day delete all of this and no longer
-    // be dependent on the URL hash to do the redirect.
-    if (hash && asPath.startsWith('/rest')) {
-      setLoadClientsideRedirectExceptions(true)
-    }
-  }, [])
 
   // Scrollable code blocks in our REST API docs and elsewhere aren't accessible
   // via keyboard navigation without setting tabindex="0".  But we don't want to set
@@ -74,8 +39,9 @@ export const RestReferencePage = ({ restOperations }: StructuredContentT) => {
     <DefaultLayout>
       {/* Doesn't matter *where* this is included because it will
       never render anything. It always just return null. */}
-      {loadClientsideRedirectExceptions && <ClientSideRedirectExceptions />}
+      <ClientSideRedirects />
       <ClientSideHighlight />
+
       <div
         className={cx(styles.restOperation, 'px-3 px-md-6 my-4 container-xl')}
         data-search="article-body"
